requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.38.0
chromadb==1.3.5
PyPDF2==3.0.1
python-multipart==0.0.6
flake8==7.3.0
pytest==9.0.1
google-genai==1.52.0